+++ Gears_Headz/src/org/usfirst/frc/team3630/robot/Robot.java
@@ -23,7 +23,7 @@ public class Robot extends IterativeRobot {
 	DriveTrain driveTrain;
 	public static AHRS ahrs;
 	public NavX myNavX;
-	public Whinch climber;
+	public Winch climber;
 	Timer autoTimer;
 	DigitalInput springEngaged;
 	// auto
@@ -50,7 +50,7 @@ public class Robot extends IterativeRobot {
 		myNavX = new NavX();
 		autoTimer = new Timer();
 		ahrs = new AHRS(SPI.Port.kMXP);
-		climber = new Whinch();
+		climber = new Winch();
 	}
 
 	/**
@@ -335,7 +335,7 @@ public class Robot extends IterativeRobot {
 		SmartDashboard.putNumber("drivetrain kI", Consts.wheelKI);
 		SmartDashboard.putNumber("drivetrain kD", Consts.wheelKD);
 		driveTrain.teleopInit();
-		climber.whinchPeriodic();
+		climber.winchPeriodic();
 	}
 
 	@Override
